Separa a insercao e a busca na tabela hash em menus separados

+++ src/Objetos/Player.h
@@ -11,7 +11,7 @@ using namespace std;
 
 class Player{
     private:
-        int id;
+        long long id;
         string pais;
         string dataDeCriacao;
         vector<Jogo> jogos;
